+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.2.0 (unreleased)
+==================
+
+* Bugfix: Default fine step was too small to feel nice.
+* Feature: Added mouse wheel scrolling.
+
 0.1.0 (2021-01-08)
 ==================
 Initial release on GitHub.